Automate build and versioning.
- release new package for every merge to main
- pin all nucliadb dependencies for every package to prevent any inconsistencies
- auto increment version with build numbers
- no longer maintain in-code changelog
