+++ .github/workflows/nucliadb.yml
@@ -141,7 +141,7 @@ jobs:
         uses: benchmark-action/github-action-benchmark@v1
         if: github.event_name == 'push'
         with:
-          tool: 'pytest'
+          tool: "pytest"
           output-file-path: benchmarks.json
           external-data-json-path: ./cache/benchmark-data.json
           save-data-file: true
@@ -149,13 +149,13 @@ jobs:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           comment-on-alert: true
           fail-on-alert: true
-          alert-comment-cc-users: '@nuclia/nuclia-db'
+          alert-comment-cc-users: "@nuclia/nuclia-db"
 
       - name: Comment benchmark result based on main
         uses: benchmark-action/github-action-benchmark@v1
         if: github.event_name != 'push'
         with:
-          tool: 'pytest'
+          tool: "pytest"
           output-file-path: benchmarks.json
           external-data-json-path: ./cache/benchmark-data.json
           save-data-file: false
@@ -163,13 +163,13 @@ jobs:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           comment-on-alert: true
           fail-on-alert: true
-          alert-comment-cc-users: '@nuclia/nuclia-db'
+          alert-comment-cc-users: "@nuclia/nuclia-db"
 
       - name: Store on GH Pages
         uses: benchmark-action/github-action-benchmark@v1
         if: github.event_name == 'push'
         with:
-          tool: 'pytest'
+          tool: "pytest"
           output-file-path: benchmarks.json
           auto-push: true
           github-token: ${{ secrets.PERFORMANCE_TOKEN }}
