+++ .github/workflows/nucliadb_dataset_release.yml
@@ -1,12 +1,10 @@
 name: Release NucliaDB Dataset
 
 on:
-  push:
-    branches:
-      - main
-    paths:
-      - ".github/workflows/nucliadb_dataset_release.yml"
-      - "nucliadb_dataset/VERSION"
+  workflow_run:
+    workflows: [Release NucliaDB Protos, Release NucliaDB SDK]
+    types:
+      - completed
 
 jobs:
   build_wheels:
@@ -25,6 +23,7 @@ jobs:
 
       - name: Install requirements
         run: |
+          python bump.py --build=${{github.run_number}}
           pip install -r nucliadb_dataset/requirements.txt
           pip install wheel
