+++ .github/workflows/nucliadb_models_release.yml
@@ -1,12 +1,10 @@
 name: Release NucliaDB Models
 
 on:
-  push:
-    branches:
-      - main
-    paths:
-      - ".github/workflows/nucliadb_models_release.yml"
-      - "nucliadb_models/VERSION"
+  workflow_run:
+    workflows: [Release NucliaDB Protos]
+    types:
+      - completed
 
 jobs:
   build_wheels:
@@ -25,6 +23,7 @@ jobs:
 
       - name: Install requirements
         run: |
+          python bump.py --build=${{github.run_number}}
           pip install -r nucliadb_models/requirements.txt
           pip install wheel
