+++ .github/workflows/nucliadb_sdk_release.yml
@@ -1,12 +1,10 @@
 name: Release NucliaDB SDK
 
 on:
-  push:
-    branches:
-      - main
-    paths:
-      - ".github/workflows/nucliadb_sdk_release.yml"
-      - "nucliadb_sdk/VERSION"
+  workflow_run:
+    workflows: [Release NucliaDB Models]
+    types:
+      - completed
 
 jobs:
   build_wheels:
@@ -25,6 +23,7 @@ jobs:
 
       - name: Install requirements
         run: |
+          python bump.py --build=${{github.run_number}}
           pip install -r nucliadb_sdk/requirements.txt
           pip install wheel
