+++ .github/workflows/nucliadb_utils.yml
@@ -1,18 +1,10 @@
 name: nucliadb Utils (py)
 
 on:
-  pull_request:
-    branches:
-      - main
-    paths:
-      - ".github/workflows/nucliadb_utils.yml"
-      - "nucliadb_utils/**"
-  push:
-    branches:
-      - main
-    paths:
-      - ".github/workflows/nucliadb_utils.yml"
-      - "nucliadb_utils/**"
+  workflow_run:
+    workflows: [Release NucliaDB Protos]
+    types:
+      - completed
 
 jobs:
   # Job to run pre-checks
