--- nucliadb/CHANGELOG
@@ -1,85 +1,0 @@
-# CHANGELOG
-
-## 2.2.5
-
-- Bump nucliadb_models
-
-## 2.2.4
-
-- Send kbid to node on shard creation
-
-## 2.2.2
-
-- Set default log level on NucliaDB to INFO
-
-## 2.2.1
-
-- Allow to start nucliadb without defining stored folders
-
-## 2.2.0
-
-- Add cancelled_by_user to TokenSplit in API
-
-## 2.0.16
-
-- Fix bug in train service returning entity positions
-
-## 2.0.15
-
-- Fix bug on text highlighting
-
-## 2.0.14
-
-- Revert returning classification labels on resource list.
-
-## 2.0.13
-
-- Include processing queue type info on resource
-
-## 2.0.12
-
-- Return classification labels on search results and resource list
-
-## 2.0.11
-
-- Fix RuntimeErrors on client disconnect
-
-## 2.0.10
-
-- Fix local shard delete resource
-
-## 2.0.9
-
-- Be resilient if proxy cannot provide an account_seq
-
-## 2.0.8
-
-- Consume account_seq from push instead from pull
-
-## 2.0.7
-
-- Be able to reindex a resource and its vectors
-
-## 2.0.6
-
-- Extend BrokerMessage to include account queue sequence
-- Include last account sequence on resource results
-
-## 2.0.5
-
-- Fix transactional error with TiKV when trying to get a nonexistent
-  KB by slug
-
-## 2.0.4
-
-- Extend OpenAPI autogenerated docs and improve endpoints
-  documentation
-
-## 2.0.3
-
-- Remove strong pinning
-
-## 2.0.2
-
-- Merge all subpackaged inside nucliadb
-- Initial release
