--- nucliadb/VERSION
@@ -1,1 +1,0 @@
-2.2.5
