+++ nucliadb/setup.py
@@ -3,11 +3,17 @@ import re
 
 from setuptools import find_packages, setup
 
+VERSION = open("../VERSION").read().strip()
+README = open("README.md").read()
+
 
 def load_reqs(filename):
     with open(filename) as reqs_file:
         return [
-            line
+            # pin nucliadb-xxx to the same version as nucliadb
+            line.strip() + f"=={VERSION}"
+            if line.startswith("nucliadb-") and "=" not in line
+            else line.strip()
             for line in reqs_file.readlines()
             if not (
                 re.match(r"\s*#", line) or re.match("-e", line) or re.match("-r", line)
@@ -19,8 +25,8 @@ requirements = load_reqs("requirements.txt")
 
 setup(
     name="nucliadb",
-    version=open("VERSION").read().strip(),
-    long_description=(open("README.md").read() + "\n" + open("CHANGELOG").read()),
+    version=VERSION,
+    long_description=README,
     long_description_content_type="text/markdown",
     classifiers=[
         "Development Status :: 4 - Beta",
