--- nucliadb_client/CHANGELOG.md
@@ -1,73 +1,0 @@
-# nucliadb_client
-
-## 2.0.0
-
-- Release
-
-## 2.0.0a7
-
-- Simulate writer and processing messages on resource commit while importing
-
-## 2.0.0a6
-
-- Fix import command
-
-## 2.0.0a5
-
-- Store GRPC channel so we can close later on
-
-## 2.0.0a4
-
-- Import/Export with binary data
-
-## 2.0.0a3
-
-- Improve export CLI argument descriptions
-
-## 2.0.0a2
-
-- Make sure slug is set on resources while importing
-
-## 2.0.0a1
-
-- Add telemetry to client's writer grpc stub
-
-## 1.1.4
-
-- Be able to get shard data from a knowledgebox
-
-## 1.1.3
-
-- Make timeout on resource.reindex configurable via method args
-
-## 1.1.2
-
-- Be able to reindex vectors too
-
-## 1.1.1
-
-- Added kb.counters()
-
-## 1.1.0
-
-- Refactor Resource.get
-
-## 1.0.1
-
-- Fix writer and manager http clients on resource and kb
-
-## 1.0.0
-
-- Export values optional
-
-## 1.0.0a2
-
-- Fix protobuf dependency
-
-## 1.0.0a1
-
-- Import and export
-
-## 0.0.1
-
-- Initial version
