--- nucliadb_client/VERSION
@@ -1,2 +1,0 @@
-2.0.0
-
