+++ nucliadb_client/requirements.txt
@@ -5,4 +5,4 @@ grpcio>=1.44.0
 aiofiles
 nucliadb-protos
 nucliadb-models
-nucliadb_telemetry
+nucliadb-telemetry>=1.2.0
