--- nucliadb_dataset/CHANGELOG.md
@@ -1,46 +1,0 @@
-# CHANGELOG
-
-## 1.2.3
-
-- Improve signatures for Datascience use case
-
-## 1.2.2
-
-- Fix headers and payload in export module
-
-## 1.2.1
-
-- Fix streamer base url
-
-## 1.2.0
-
-- Adapt to nucliadb_sdk refactor
-
-## 1.1.4
-
-- Add Cloud datasets implementation
-
-## 1.1.3
-
-- Fix sentence classification texts
-
-## 1.1.2
-
-- Fix sentence classification schema
-
-## 1.1.1
-
-- Add `download_all_partitions` generic function to make it easier
-
-## 1.1.0
-
-- Implement a common interface for different dataset sources
-- Refactor NucliaDBDataset to match the new interface
-
-## 1.0.1
-
-- Fix storing path environment variable
-
-## 1.0.0
-
-- Initial release
