--- nucliadb_dataset/VERSION
@@ -1,1 +1,0 @@
-1.2.3
