+++ nucliadb_dataset/requirements.txt
@@ -1,12 +1,10 @@
 protobuf
 types-protobuf
 grpcio
-nucliadb-protos
 requests
 aiohttp
 pydantic_argparse
 pyarrow
-nucliadb_sdk>=1.2.0
 boto3
 google-auth
 google-cloud-storage
@@ -14,3 +12,5 @@ boto3-stubs
 google-auth-stubs
 types-google-cloud-ndb
 types-boto3
+nucliadb-protos
+nucliadb-sdk
