--- nucliadb_models/CHANGELOG.rst
@@ -1,31 +1,0 @@
-CHANGELOG
-=========
-
-2.0.5
------
-- Add relations to extracted field metadata
-
-2.0.4
------
-- Add cancelled_by_user to TokenSplit
-
-2.0.3
------
-
-- Add error metadata on fields' FieldData
-
-2.0.2
------
-
-- Upgrade models with latest version release for nucliadb_dataset
-
-2.0.1
------
-
-- Remove labels from Sentence in search results
-- Renamed ParagraphPosition to TextPosition
-
-1.0.0
-------
-
-- Initial release
