--- nucliadb_models/VERSION
@@ -1,1 +1,0 @@
-2.0.6
