+++ nucliadb_models/requirements.txt
@@ -1,2 +1,2 @@
 pydantic>=1.8.2
-nucliadb_protos>=0.0.36
+nucliadb-protos
