--- nucliadb_node/CHANGELOG
@@ -1,7 +1,0 @@
-CHANGELOG
-=========
-
-1.0.0 (unreleased)
----------------------
-
-- Initial release .
