--- nucliadb_node/VERSION
@@ -1,1 +1,0 @@
-1.0.0
