--- nucliadb_protos/python/VERSION
@@ -1,1 +1,0 @@
-0.0.51
