--- nucliadb_sdk/CHANGELOG.md
@@ -1,50 +1,0 @@
-# nucliadb_sdk
-
-## 1.2.7
-
-- Cloud as the default option when creating a client
-
-## 1.2.6
-
-- Allow to delete a KB on local deployment
-
-## 1.2.5
-
-- Fix get_or_create
-
-## 1.2.4
-
-- Improve signatures for Datascience use case
-
-## 1.2.3
-
-- Simplify creation of KBs and vector indexing
-
-## 1.2.2
-
-- Allow to define the min_score on semantic search via SDK
-
-## 1.2.1
-
-- Upload testing max GRPC size for importing via GRPC
-- Fix typo on `length` method
-
-## 1.2.0
-
-- Refactor Knowledgebox to accept only an initialized sdk NucliaDBClient
-
-## 1.1.2
-
-- Add nucliadb_models requirements
-
-## 1.1.1
-
-- Add on client the option to have component based urls
-
-## 1.1.0
-
-- Suport of auto creation of vectorsets, easier interface.
-
-## 1.0.0
-
-- Initial Release
