+++ nucliadb_sdk/requirements.txt
@@ -1,2 +1,2 @@
 httpx
-nucliadb_models>=2.0.2
+nucliadb-models
