+++ nucliadb_telemetry/VERSION
@@ -1,1 +1,1 @@
-1.1.9
+1.2.0
