+++ nucliadb_telemetry/setup.py
@@ -2,38 +2,33 @@ import re
 
 from setuptools import find_packages, setup  # type: ignore
 
-with open("README.md") as f:
-    readme = f.read()
+VERSION = open("VERSION").read().strip()
+README = open("README.md").read()
 
 
 def load_reqs(filename):
     with open(filename) as reqs_file:
         return [
-            line
+            # pin nucliadb-xxx to the same version as nucliadb
+            line.strip() + f"=={VERSION}"
+            if line.startswith("nucliadb-") and "=" not in line
+            else line.strip()
             for line in reqs_file.readlines()
             if not (
-                re.match(r"\s*#", line)  # noqa
-                or re.match("-e", line)
-                or re.match("-r", line)
+                re.match(r"\s*#", line) or re.match("-e", line) or re.match("-r", line)
             )
         ]
-
-
-def load_version():
-    with open("VERSION", "r") as f:
-        version = f.read().strip()
-    return version
 
 
 requirements = load_reqs("requirements.txt")
 
 setup(
     name="nucliadb_telemetry",
-    version=load_version(),
+    version=VERSION,
     author="nucliadb Authors",
     author_email="nucliadb@nuclia.com",
     description="NucliaDB Telemetry Library Python process",
-    long_description=readme,
+    long_description=README,
     long_description_content_type="text/markdown",
     license="MIT",
     url="https://github.com/nuclia/nucliadb",
