--- nucliadb_utils/CHANGELOG.rst
@@ -1,100 +1,0 @@
-CHANGELOG
-=========
-
-1.0.18
-------
-
-- Audit improvements
-
-1.0.17
-------
-
-- Various storage fixes
-
-1.0.15
-------
-
-- Fix delete reindex payload
-
-1.0.14
-------
-
-- Add temporary logging
-
-1.0.13
-------
-
-- Make sure starlette is running on version >= 0.21.0 as it fixes a RuntimeError on client disconnect
-
-1.0.12
-------
-
-- Fix transaction pubsub keyerror
-
-1.0.11
-------
-- Add OpenAPI utils for fastapi doc generation
-
-1.0.10
-------
-
-- Relax dependencies
-
-1.0.9
------
-
-- Parametrize upload token expiration
-
-
-1.0.8 (2022-07-20)
-------------------
-
-- Upgrade aiohttp
-
-
-1.0.7 (2022-07-08)
-------------------
-
-- Updated nats.py package
-
-
-1.0.6 (2022-05-31)
-------------------
-
-- Split running settings on a different module
-
-
-1.0.5 (2022-05-31)
-------------------
-
-- Nothing changed yet.
-
-
-1.0.4 (2022-05-31)
-------------------
-
-- Nothing changed yet.
-
-
-1.0.3 (2022-05-31)
-------------------
-
-- Split dependencies for easier installation
-
-
-1.0.2 (2022-05-31)
-------------------
-
-- Update Pydantic dependency to >= 1.8.2
-
-
-1.0.1 (2022-05-31)
-------------------
-
-- Fix release
-
-
-1.0.0 (2022-05-31)
-------------------
-
-- Initial release
