+++ nucliadb_utils/nucliadb_utils/fastapi/exclude_urls.py
@@ -16,17 +16,7 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
+from nucliadb_telemetry.fastapi import ExcludeList
 
-from typing import Iterable
-from urllib.parse import urlparse
-
-
-class ExcludeList:
-
-    """Class to exclude certain paths (given as a list of regexes) from tracing requests"""
-
-    def __init__(self, excluded_urls: Iterable[str]):
-        self._excluded_urls = excluded_urls
-
-    def url_disabled(self, url: str) -> bool:
-        return bool(self._excluded_urls and urlparse(url).path in self._excluded_urls)
+# backwards compatibility, deprecated -- use nucliadb_telemetry.fastapi.ExcludeList
+__all__ = ("ExcludeList",)
