+++ nucliadb_utils/requirements-storages.txt
@@ -2,7 +2,7 @@ oauth2client>=4.1.3
 aiobotocore>=2.1.1
 google-api-python-client>=2.37.0
 opentelemetry-instrumentation-aiohttp-client
-nucliadb_telemetry
+nucliadb-telemetry>=1.2.0
 types-aiofiles>=0.8.3
 aiofiles>=0.8.0
 backoff>=1.11.1
